MAP-1406 package refactoring and name simplification
