+++ src/test/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/LocationCapacityResourceIntTest.kt
@@ -17,7 +17,7 @@ import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.SpecialistCellType
 import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.LocationRepository
 import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.buildCell
 import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.buildResidentialLocation
-import uk.gov.justice.digital.hmpps.locationsinsideprison.resource.locationresource.EXPECTED_USERNAME
+import uk.gov.justice.digital.hmpps.locationsinsideprison.resource.location.EXPECTED_USERNAME
 import uk.gov.justice.hmpps.test.kotlin.auth.WithMockAuthUser
 import java.time.Clock
 import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.ResidentialLocation as ResidentialLocationJPA
