+++ src/test/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/location/LocationKeyTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.locationsinsideprison.resource.locationresource
+package uk.gov.justice.digital.hmpps.locationsinsideprison.resource.location
 
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
@@ -7,7 +7,7 @@ import uk.gov.justice.digital.hmpps.locationsinsideprison.integration.CommonData
 import uk.gov.justice.hmpps.test.kotlin.auth.WithMockAuthUser
 
 @WithMockAuthUser(username = EXPECTED_USERNAME)
-class LocationResourceKeyTest : CommonDataTestBase() {
+class LocationKeyTest : CommonDataTestBase() {
 
   @DisplayName("GET /locations/key/{key}")
   @Nested
