+++ src/test/kotlin/uk/gov/justice/digital/hmpps/locationsinsideprison/resource/location/LocationTransformTest.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.locationsinsideprison.resource.locationresource
+package uk.gov.justice.digital.hmpps.locationsinsideprison.resource.location
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
@@ -14,7 +14,7 @@ import uk.gov.justice.digital.hmpps.locationsinsideprison.jpa.repository.buildCe
 import uk.gov.justice.hmpps.test.kotlin.auth.WithMockAuthUser
 
 @WithMockAuthUser(username = EXPECTED_USERNAME)
-class LocationResourceTransformTest : CommonDataTestBase() {
+class LocationTransformTest : CommonDataTestBase() {
 
   @DisplayName("PUT /locations/{id}/used-for-type")
   @Nested
